Add more details to Connection response error message (#339)
The error message produced by an unsuccessful request made by the `Connection` class is somewhat opaque and lacks context for effectively diagnosing and/or debugging the problem. This extends the error message produced by failed `Connection` API requests with the contents of the `meta` field (if present) in the error response body.