Fix shapes of weights and means, properly initialize weights
Fixed the shapes of the weights and means object by updating the tuples in w_shape and m_shape. Initialized and normalized the weights appropriately in the reset method of BosonicModes. Temporarily changed the initialization of BaseBosonicState to go through only in the weight-one (single gaussian) case before the indices are updated to the general case.