Update BosonicBackend with methods for preparing and acting on the BosonicModes circuit
Add method run_prog to BosonicBackend. This method accepts an sf.Program object from bosonic engine to traverse it, properly instantiate the BosonicModes circuit with objects of the right shape, and then to apply operations to the circuit. For now the operation application is done exactly as in LocalEngine. Circuit initialization is done in init_circuit: this method goes through the Preparation classes it sees in the program and asks for weights, means and covariance objects, either stored in the parameters of the preparation or generated in the prepare_ methods in the backend itself. Empty prepare methods also added in anticipation of completed methods from another branch.