Gkps mbsqueezing (#509)
* first commit

* Adds generaldyne sampling and postselection

- adds functionality to sample generaldyne measurements from bosonic modes
- adds functionality to post-select (i.e. conditionally update) the remaining bosonic modes
- homodyne and heterodyne are implemented as subcases of general-dyne

* Fixed covariance conventions

Covariances do not assume a specific value of hbar, and diagonals match quadrature variances.

* Update LocalEngine, BosonicEngine in engine.py

Change the __new__ method in LocalEngine to have correct behaviour depending on whether the specified backend is bosonic. In BosonicEngine.run include some code from Local Engine required for operation. In BosonicEngine._run_program make sure 'batches' is fed into the backend method and all outputs are collected from the backend.

* Properly format some backend files

Run black -l 100 on __init__, base, and states in \backends and ops in \bosonicbackend.

* Measurement based-squeezing

- Adds measurement-based squeezing to SF ops.py file
- Adds measurement-based squeezing to bosonic backend.py file
- Adds measurement-based squeezing to bosoniccircuit.py file
- Adds/corrects relevant functions in the bosoniccircuit.py file (loss, expanding Gaussian CPTP matrices)
- Sampling homodyne values for single shot mb squeezing requires will require updates being developed in the gkps-sampling branch

* Add to allowed preparations in the bosonic compiler

Update the bosonic compiler class in compilers\bosonic.py to allow for Fock, Ket, DensityMatrix, Bosonic, GKP, Cat, and Comb preparation classes.

* Import numpy as np in bosonic.py

Import numpy as np in the bosonic backend module rather than specific objects from numpy because of the growing list of them.

* Update BosonicBackend with methods for preparing and acting on the BosonicModes circuit

Add method run_prog to BosonicBackend. This method accepts an sf.Program object from bosonic engine to traverse it, properly instantiate the BosonicModes circuit with objects of the right shape, and then to apply operations to the circuit. For now the operation application is done exactly as in LocalEngine. Circuit initialization is done in init_circuit: this method goes through the Preparation classes it sees in the program and asks for weights, means and covariance objects, either stored in the parameters of the preparation or generated in the prepare_ methods in the backend itself. Empty prepare methods also added in anticipation of completed methods from another branch.

* Some empty preparation classes

Prune GKP and add Comb as empty preparation classes in anticipation of completed classes from another branch.

* Small updates to bosoniccircuit.py

Change num_subsystems to default to 1 in BosonicModes.__init__ and prevent circuit from resetting to vacuum by default. Also temporarily prune docstring of BosonicModes and format with black -l 100.

* mbsqueezing angle

- Matches convention of complex squeezing for mbsqueezing to squeezing.

* Fixes apply_channel

fixes update_means function in apply_channel

* update requirements

includes the walrus

* Update requirements

Includes the walrus

* Fix sphinx error by adding thewalrus requirement

Add thewalrus requirement to doc\requirements.txt to fix the failed sphinx test on pull request,

* Formatting with black

Formatting with black

* Formatted with black

* Change Cat to Catstate in BosonicBackend

* Comment out decompositions in the bosonic compiler

Attempt to fix the sphinx error on pull request by commenting out the decompositions in the bosonic compiler.

* Updating single shot measurement-based squeezing

Updates mbsqueeze function

* Add Bosonic, GKP, Comb classes to lists for unit tests and docs

Add Bosonic, GKP, and Comb classes to the state preparation class list in ops.py for unit tests and docs.

* Update requirements.txt in \doc to match those in setup.py

Update versions of packages to match the requirements files in the root directory.

* Revert "Update requirements.txt in \doc to match those in setup.py"

This reverts commit bddbe1694aa02e01a6af714ceec8cfec8edfc8bd.

* Minor changes to bosonicbackend\backend.py and compilers\bosonic.py

backend.py:
- Change importing line to import specifics ops
- Add longer docstrings to to_xp and from_xp
- Make sure all empty preparations return nothing

bosonic.py:
- remove MeasureHeterodyne from compiler primitives.

* Remove references to bosonic backend in conftest.py

Remove references to bosonic backend in conftest.py to see if this makes the checks pass.

* Add "pass" to GKP and Comb classes in ops.py

* Move ops import command to init_circuit in BosonicBackend

* Change Cat to Catstate in the bosonic compiler

* Important updates to run_prog in bosonic backend

Update run_prog in BosonicBackend so that:
- Things are properly nested in the for loop (empty lists and dicts placed and the return command are placed outside the loop)
- It ignores Preparation classes while it is applying operations to the circuit (as modes have already been initialized based on the preparations in init_circuit). NOTE: this produces the wrong behaviour in case of non-initial preparations, which we can address later.

* Run backend.py through black

* Get run_prog to properly apply gaussian preparation

Since init_circuit initializes gaussian preparations in a vacuum, we would like run_prog to properly apply the corresponding gaussian transformation when it encounters a gaussian preparation. This is what this change does. In the future, we can modify init_circuit to implement gaussian prep fully.

* Fix weights shape in init_circuit (denest list)

* Update add_mode function in BosonicModes

Change add_mode to accept a list of numbers of modes per peak and remove references to self._trunc.

* Minor changes for mb squeezing

- add mbSgate to list of gates in compiler
-ensure weights objects are arrays for easier manipulation

* Front-end access to samples from mb gates

When a user applies a single-shot ancilla-assisted gate to a mode, a measurement outcome on the ancilla is associated with the mode.

This change creates and populates a dictionary ancilla_samples in the Results class so that users can access the measurement outcomes from each ancilla-assisted gate.

* Tidying mb squeezing

- Making sure hbar conventions are respected
- Deleting extra info from the covs and means once an ancilla is consumed since it just resets to vacuum.

* Ran black

Ran code through black

* Capitalizing class name

mbSgate -> MbSgate

* mbSgate -> MbSgate

* mbSgate -> MbSgate

* Style change

Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>

* numpy dtype update

Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>

* Style change

Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>

* Style change

Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>

* Style change

Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>

* Style change

Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>

* Style change

Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>

* Style change

Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>

* Style change

Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>

Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: ilan-tz <57886357+ilan-tz@users.noreply.github.com>
Co-authored-by: Nicolas Quesada <zeitus@gmail.com>
Co-authored-by: GDauphinais <48553380+GDauphinais@users.noreply.github.com>