Gkps state prep gkp complexcat fock (#516)
* first commit

* Added incomplete methods for preparing non-gaussian states

* Update LocalEngine, BosonicEngine in engine.py

Change the __new__ method in LocalEngine to have correct behaviour depending on whether the specified backend is bosonic. In BosonicEngine.run include some code from Local Engine required for operation. In BosonicEngine._run_program make sure 'batches' is fed into the backend method and all outputs are collected from the backend.

* Properly format some backend files

Run black -l 100 on __init__, base, and states in \backends and ops in \bosonicbackend.

* Add to allowed preparations in the bosonic compiler

Update the bosonic compiler class in compilers\bosonic.py to allow for Fock, Ket, DensityMatrix, Bosonic, GKP, Cat, and Comb preparation classes.

* Import numpy as np in bosonic.py

Import numpy as np in the bosonic backend module rather than specific objects from numpy because of the growing list of them.

* Update BosonicBackend with methods for preparing and acting on the BosonicModes circuit

Add method run_prog to BosonicBackend. This method accepts an sf.Program object from bosonic engine to traverse it, properly instantiate the BosonicModes circuit with objects of the right shape, and then to apply operations to the circuit. For now the operation application is done exactly as in LocalEngine. Circuit initialization is done in init_circuit: this method goes through the Preparation classes it sees in the program and asks for weights, means and covariance objects, either stored in the parameters of the preparation or generated in the prepare_ methods in the backend itself. Empty prepare methods also added in anticipation of completed methods from another branch.

* Some empty preparation classes

Prune GKP and add Comb as empty preparation classes in anticipation of completed classes from another branch.

* Small updates to bosoniccircuit.py

Change num_subsystems to default to 1 in BosonicModes.__init__ and prevent circuit from resetting to vacuum by default. Also temporarily prune docstring of BosonicModes and format with black -l 100.

* Fix sphinx error by adding thewalrus requirement

Add thewalrus requirement to doc\requirements.txt to fix the failed sphinx test on pull request,

* Change Cat to Catstate in BosonicBackend

* Comment out decompositions in the bosonic compiler

Attempt to fix the sphinx error on pull request by commenting out the decompositions in the bosonic compiler.

* Add Bosonic, GKP, Comb classes to lists for unit tests and docs

Add Bosonic, GKP, and Comb classes to the state preparation class list in ops.py for unit tests and docs.

* Update requirements.txt in \doc to match those in setup.py

Update versions of packages to match the requirements files in the root directory.

* Revert "Update requirements.txt in \doc to match those in setup.py"

This reverts commit bddbe1694aa02e01a6af714ceec8cfec8edfc8bd.

* added prepare_gkp method in backend.py

* ran black

* Minor changes to bosonicbackend\backend.py and compilers\bosonic.py

backend.py:
- Change importing line to import specifics ops
- Add longer docstrings to to_xp and from_xp
- Make sure all empty preparations return nothing

bosonic.py:
- remove MeasureHeterodyne from compiler primitives.

* fixed a missing hbar in method prepare_gkp

* fixed missing normalization for GKP weights

* Remove references to bosonic backend in conftest.py

Remove references to bosonic backend in conftest.py to see if this makes the checks pass.

* Add "pass" to GKP and Comb classes in ops.py

* Move ops import command to init_circuit in BosonicBackend

* Change Cat to Catstate in the bosonic compiler

* Important updates to run_prog in bosonic backend

Update run_prog in BosonicBackend so that:
- Things are properly nested in the for loop (empty lists and dicts placed and the return command are placed outside the loop)
- It ignores Preparation classes while it is applying operations to the circuit (as modes have already been initialized based on the preparations in init_circuit). NOTE: this produces the wrong behaviour in case of non-initial preparations, which we can address later.

* Run backend.py through black

* Added a subcase of prepare_cat, minor modifications to prepare_gkp

* fixes merge conflict

* Adds the functions for Fock state preparation (#506)

* Adds the functions

* Passes black

* dummy change

* dummy change

* resolves issues in merging

* removes useless comment

* Get run_prog to properly apply gaussian preparation

Since init_circuit initializes gaussian preparations in a vacuum, we would like run_prog to properly apply the corresponding gaussian transformation when it encounters a gaussian preparation. This is what this change does. In the future, we can modify init_circuit to implement gaussian prep fully.

* Fix weights shape in init_circuit (denest list)

* Finished adding the real Cat state preparation

* added a hbar factor in means of prepare_gkp

* Removed norm from numpy import

* added isin import from numpy

* fixed self.hbar

* few typos corrected in backend.py, modified __init__ method for classes Catstate and GKP

* minor bugs preventing the usage of non-Gaussian states have been corrected

* Fixed bugs in prepare_cat method of backedn.py and switched to import numpy, modified ops.py to properly initialize parameters

* ran Black

* ran proper black version

* Fixed a factor of pi mistake in class Catstate in ops.py

* Fixes hbar in prepare cat

* Adds if branching when preparing Fock states

* simplifies logic

* simplifies logic

* removes unnecesary if statement

* passes black

* fixes gkp weight cutoff

* Modified prepare_cat function: D is now a parameter that can be used by user. Fixed a few issues in prepare_gkp from pr.

* ran black

* changes dtype of means

* adding factor of hbar

* Removes real cat representation

* Minor changes

- change front end call to only require alpha and p
- vacuum cov factor of hbar

* Apply suggestions from code review

Co-authored-by: Nicolas Quesada <zeitus@gmail.com>

* minor typo fix

* black formatting

Co-authored-by: Josh Izaac <josh146@gmail.com>
Co-authored-by: Guillaume Dauphinais <guillaume@xanadu.ai>
Co-authored-by: ilan-tz <57886357+ilan-tz@users.noreply.github.com>
Co-authored-by: Nicolas Quesada <zeitus@gmail.com>