fix [test_bosonic_backend] - test_cat_complex
To check for real [or complex] output the following assertion was used:
`assert [not] np.allclose(state.weights().real, state.weights())`
Notice however that this is prune to numerical failure as it relies on rounding. Numpy has implemented the functions `np.isreal`, `np.iscomplex` functions to check whether the input array contains real or complex values.

Care should be taken when evaluating for complex values: `np.iscomplex` returns false for numbers with zero imaginary part (i.e., 2+1j*0.00000000001). If we rather evluate that the dtype of the array is `complex`, si better to use `np.iscomplexobj`.