Update plot.py - Show plots when not running in a notebook (using try and except blocks)
The code works fine when using it in a notebook but the graph does not show when using other IDEs such as Spyder. These lines are added in try and except blocks to make sure that when it is not running on a notebook, the graph will be displayed in a browser.