Update plot.py - Show plots when not running in a notebook (#669)
* Update plot.py - Show plots when not running in a notebook

The code works fine when using it in a notebook but the graph does not show when using other IDEs such as Spyder. The lines are added to make sure that when it is not running on a notebook, the graph will be displayed in a browser.

* Update plot.py - Show plots when not running in a notebook (using try and except blocks)

The code works fine when using it in a notebook but the graph does not show when using other IDEs such as Spyder. These lines are added in try and except blocks to make sure that when it is not running on a notebook, the graph will be displayed in a browser.

* Update plot.py - Show plots when not running in a notebook (using try and except blocks)

The code works fine when using it in a notebook but the graph does not show when using other IDEs such as Spyder. These lines are added in try and except blocks to make sure that when it is not running on a notebook, the graph will be displayed in a browser.

* Update Plot.py

remove the trailing whitespace

* Update plot.py

Remove trailing whitespace.

* Update plot.py 

Updated to show the graph when not running as a notebook. Black autoformatted.

* Updates plot.py with black ran locally.

* Updates plot.py- Show plots when not running in a notebook

* Adds the contribution regarding the PR #669

* Update .github/CHANGELOG.md

uses suggested wordings

Co-authored-by: Theodor <theodor.isacsson@gmail.com>

* suggestions applied

* adds the removed spaces again

* make format after adding spaces back

Co-authored-by: Theodor <theodor@xanadu.ai>
Co-authored-by: Hossein Seifoory <hseifoory@nvidia.com>
Co-authored-by: Theodor <theodor.isacsson@gmail.com>